+++ src/Products/PloneMeeting/browser/views.py
@@ -2200,13 +2200,14 @@ class DisplayMeetingItemVoters(BrowserView):
                     i = vote_number
                     linked_numbers = _get_linked_item_vote_numbers(
                         item, self.context, vote_number=vote_number)
+                    # aggregate linked votes to original one
                     if linked_numbers:
-                        i = linked_numbers[0]
+                        i = min(linked_numbers)
                     if i not in data:
                         data[i] = 0
                     vote_count = item.getVoteCount('any_voted', vote_number=vote_number)
                     data[i] += vote_count
-                # now if we have a element in res < total_voters, we miss some votes
+                # now if we have an element in res < total_voters, we miss some votes
                 for count in data.values():
                     if count < total_voters:
                         res['secret'].append(item)